requirements copy.txt
asgiref==3.5.0
Django==4.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-phonenumber-field==6.0.0
phonenumbers==8.12.38
Pillow==9.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
